% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{syrian_places}
\alias{syrian_places}
\title{Latitude and Longitude Data for Locations in Syria}
\format{A data frame with columns:
\describe{
 \item{City}{Name of City.}
 \item{Neighborhood}{Name of area.}
 \item{Latitude}{Latitude of location.}
 \item{Longitude}{Longitude of location.}
}}
\source{
Google's geocoding API
}
\usage{
syrian_places
}
\description{
Latitude and Longitude Data for Locations in Syria
}
\examples{
 syrian_places
}
\keyword{datasets}

