% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syrianarchive.R
\docType{package}
\name{syrianarchive}
\alias{syrianarchive}
\alias{syrianarchive-package}
\title{The Syrian Archive}
\description{
Data scraped from \url{https://syrianarchive.org/}
}
\seealso{
\code{\link{violations}}
}

